# TodoList/TodoList.Server/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TodoList.Server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/todoitems": {
      "get": {
        "tags": [
          "TodoList.Server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "TodoList.Server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TodoItemDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/todoitems/complete": {
      "get": {
        "tags": [
          "TodoList.Server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/todoitems/{id}": {
      "get": {
        "tags": [
          "TodoList.Server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "TodoList.Server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TodoItemDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "TodoList.Server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "TodoItemDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "isComplete": {
          "type": "boolean"
        },
        "name": {
          "type": "string"
        },
        "dateAndTime": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}